Reorganised generic popup dialog
